Day 13 - Part 2 WIP Refactor - Modules for Valley & Pattern - Valley::summarise_notes() - Pattern::find_mirror() -> Iter - find_horizontal_mirror_max/min() - find_vertical_mirror_max/min()
