+++ src/bin/day13/sample_p2.txt
@@ -1,0 +1,15 @@
+..##..##.
+..#.##.#.
+##......#
+##......#
+..#.##.#.
+..##..##.
+#.#.##.#.
+
+#....#..#
+#....#..#
+..##..###
+#####.##.
+#####.##.
+..##..###
+#....#..#
